Add Abyss support
