+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "theresmore-automation",
-  "version": "4.11.0",
+  "version": "4.11.1",
   "description": "Automation suite for Theresmore game",
   "keywords": [
     "userscript",
